12.19前后端合并

+++ 04.SourceCode/WPFMVVMDemo-master/Operation/BasicOpreation.cs
@@ -10,21 +10,21 @@ namespace Operation
     //二元运算
     public class Basic_Opreation
     {
-        public Double Add(string num1, string num2)
+        public Decimal Add(string num1, string num2)
         {
-            return Convert.ToDouble(num1) + Convert.ToDouble(num2);
+            return Convert.ToDecimal(num1) + Convert.ToDecimal(num2);
 
         }
-        public Double Sub(string num1, string num2)
+        public Decimal Sub(string num1, string num2)
         {
             
-            return Convert.ToDouble(num1) - Convert.ToDouble(num2);
+            return Convert.ToDecimal(num1) - Convert.ToDecimal(num2);
         }
-        public Double Mul(string num1, string num2)
+        public Decimal Mul(string num1, string num2)
         {
-            return Convert.ToDouble(num1) * Convert.ToDouble(num2);
+            return Convert.ToDecimal(num1) * Convert.ToDecimal(num2);
         }
-        public Double Div(string num1, string num2)
+        public Decimal Div(string num1, string num2)
         {
             
 
@@ -33,7 +33,7 @@ namespace Operation
                 Cache.topCache = "除数不能为零";
 
             }
-                return Convert.ToDouble(num1) / Convert.ToDouble(num2);
+                return Convert.ToDecimal(num1) / Convert.ToDecimal(num2);
 
             
         }
